backend/requirements.txt
python-dotenv==1.0.0
langchain==0.1.17
langchain-google-genai==0.0.12
langchain-openai==0.1.1
langchain-community==0.0.33
pydantic==2.5.3
pydantic-settings==2.1.0
groq==0.4.2
aiohttp==3.9.1
asyncio-contextmanager==1.0.0
python-json-logger==2.0.7
supabase==2.10.0
pytest==8.3.5
langgraph==0.2.28
fastapi==0.116.1
uvicorn[standard]==0.35.0

